+++ packages/commerce-sdk-react/src/auth/index.ts
@@ -53,6 +53,7 @@ interface AuthConfig extends ApiClientConfigParams {
     passwordlessLoginCallbackURI?: string
     refreshTokenRegisteredCookieTTL?: number
     refreshTokenGuestCookieTTL?: number
+    hybridAuthEnabled?: boolean
 }
 
 interface JWTHeaders {
@@ -243,6 +244,8 @@ class Auth {
         | ((loginId: string, usid: string, refresh: boolean) => Promise<TokenResponse>)
         | undefined
 
+    private hybridAuthEnabled: boolean
+
     constructor(config: AuthConfig) {
         // Special endpoint for injecting SLAS private client secret.
         const baseUrl = config.proxy.split(MOBIFY_PATH)[0]
@@ -339,6 +342,8 @@ class Auth {
                 ? passwordlessLoginCallbackURI
                 : `${baseUrl}${passwordlessLoginCallbackURI}`
             : ''
+
+        this.hybridAuthEnabled = config.hybridAuthEnabled || false
     }
 
     get(name: AuthDataKeys) {
@@ -576,6 +581,14 @@ class Auth {
      * registered shopper refresh-token and restores session and basket on SFRA.
      */
     private clearECOMSession() {
+        /**
+         * If `hybridAuthEnabled` is true, dwsid cookie must not be cleared.
+         * This makes sure the session-bridged dwsid, received from `/oauth2/token` call on shopper login
+         * is NOT cleared and can be used to maintain the server affinity.
+         */
+        if (this.hybridAuthEnabled) {
+            return
+        }
         const {key, storageType} = DATA_MAP[DWSID_COOKIE_NAME]
         const store = this.stores[storageType]
         store.delete(key)