+++ packages/commerce-sdk-react/src/provider.tsx
@@ -50,6 +50,7 @@ export interface CommerceApiProviderProps extends ApiClientConfigParams {
     refreshTokenGuestCookieTTL?: number
     apiClients?: ApiClients
     disableAuthInit?: boolean
+    hybridAuthEnabled?: boolean
 }
 
 /**
@@ -108,6 +109,17 @@ export const AuthContext = React.createContext({} as Auth)
  * Non-PWA Kit users can enable private client mode by passing in a client secret
  * directly to the provider. However, be careful when doing this as you will have
  * to make sure the secret is not unexpectedly exposed to the client.
+ * 
+ * 
+ * `hybridAuthEnabled` is an optional flag that indicates the current Site has Hybrid Auth enabled.
+ * This drives the behavior of the `clearECOMSession` method. If `hybridAuthEnabled` is true,
+ * the `clearECOMSession` method will not be called. This makes sure the session-bridged dwsid, received from `/oauth2/token` call
+ * on shopper login is NOT cleared and can be used to maintain the server affinity.
+ * 
+ * `hybridAuthEnabled` flag can also be used to drive other Hybrid Auth specific behaviors in the future.
+ * 
+ * Note: `hybridAuthEnabled` should NOT be set to true for hybrid storefronts using Plugin SLAS as we need the dwsid to be deleted
+ * to force session-bridging on SFRA as in this case, the `oauth2/token` call does not return a dwsid.
  *
  * @returns Provider to wrap your app with
  */
@@ -134,7 +146,8 @@ const CommerceApiProvider = (props: CommerceApiProviderProps): ReactElement => {
         refreshTokenRegisteredCookieTTL,
         refreshTokenGuestCookieTTL,
         apiClients,
-        disableAuthInit = false
+        disableAuthInit = false,
+        hybridAuthEnabled = false
     } = props
 
     // Set the logger based on provided configuration, or default to the console object if no logger is provided
@@ -157,7 +170,8 @@ const CommerceApiProvider = (props: CommerceApiProviderProps): ReactElement => {
             defaultDnt,
             passwordlessLoginCallbackURI,
             refreshTokenRegisteredCookieTTL,
-            refreshTokenGuestCookieTTL
+            refreshTokenGuestCookieTTL,
+            hybridAuthEnabled
         })
     }, [
         clientId,
@@ -176,7 +190,8 @@ const CommerceApiProvider = (props: CommerceApiProviderProps): ReactElement => {
         passwordlessLoginCallbackURI,
         refreshTokenRegisteredCookieTTL,
         refreshTokenGuestCookieTTL,
-        apiClients
+        apiClients,
+        hybridAuthEnabled
     ])
 
     const dwsid = auth.get(DWSID_COOKIE_NAME)